socialnetwork/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
requests[security]==2.24.0
drf-yasg==1.17.1
djangorestframework-simplejwt==4.6.0
celery==5.0.5